refactor: moving tests files into tests folder
