+++ src/departments/tests/departments.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { DepartmentsController } from './departments.controller';
+import { DepartmentsController } from '../departments.controller';
 
 describe('DepartmentsController', () => {
   let controller: DepartmentsController;
