+++ src/users/tests/users.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { UsersService } from './users.service';
+import { UsersService } from '../users.service';
 import { PrismaService } from 'src/prisma/prisma.service';
 
 @Module({
